+++ sql_to_fnc.constans.ts
@@ -1,2 +1,3 @@
 export const privUser = 'lac_api';
 export const privOwner = 'lac_group';
+export const datePlaceholder = 'DD.MM.RRRR';