+++ src/common.ts
@@ -1,4 +1,4 @@
-import {FieldDefinition} from "./sql_to_fnc.interfaces";
+import { FieldDefinition } from "./sql_to_fnc.interfaces";
 
 export function capitalize(str: string): string {
   return str.charAt(0).toUpperCase() + str.slice(1);
@@ -12,10 +12,22 @@ export function isNumber(item: FieldDefinition): boolean {
 }
 
 export function isString(item: FieldDefinition): boolean {
-  if (
-    (item.type.toUpperCase().indexOf('VARCHAR') >= 0)
-    || (item.type.toUpperCase().indexOf('BPCHAR') >= 0)
-    || (item.type.toUpperCase().indexOf('TEXT') >= 0)
-  ) return true;
+  if (['VARCHAR', 'BPCHAR', 'TEXT'].indexOf(item.type.toUpperCase()) >= 0) {
+    return true;
+  }
+  return false;
+}
+
+export function isBoolean(item: FieldDefinition): boolean {
+  if (['BOOL'].indexOf(item.type.toUpperCase()) >= 0) {
+    return true;
+  }
+  return false;
+}
+
+export function isDate(item: FieldDefinition): boolean {
+  if (['DATE', 'TIMESTAMP'].indexOf(item.type.toUpperCase()) >= 0) {
+    return true;
+  }
   return false;
 }